news-bot/requirements.txt
newspaper3k
nltk
